Update MapMove, datas and docs

+++ Sample/Scripts/Source/NodeFunctions/Math_meta.lua
@@ -360,7 +360,7 @@ local _METADATA = {
             },
             ["return"] = {
                 "value",
-                value = "float",
+                value = "int",
             },
             Pure = true,
         },
